requirements/base.txt
python-dotenv==1.0.1
fastapi==0.115.4
loguru==0.7.2
crewai==0.70.1